% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QcPSP.R
\name{extractArtemis2009FormatFromPSPForMetaModelling}
\alias{extractArtemis2009FormatFromPSPForMetaModelling}
\title{Extract plot list for Artemis-2009 simulation}
\usage{
extractArtemis2009FormatFromPSPForMetaModelling(QcPSPData, list_ID_PE_MES)
}
\arguments{
\item{QcPSPData}{the database that is retrieved through the restoreQcPSPData
function}

\item{list_ID_PE_MES}{a vector of numerics that stand for the ID_PE_MES. The
ID_PE_MES field is found in the plotMeasurements data.frame of the
database.}
}
\value{
a data.frame object formatted for Capsis simulation
}
\description{
Extract plot list for Artemis-2009 simulation
}
