% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QcPSP.R
\name{extractArtemis2009FormatFromTSP4ForMetaModelling}
\alias{extractArtemis2009FormatFromTSP4ForMetaModelling}
\title{Extract plot list for Artemis-2009 simulation}
\usage{
extractArtemis2009FormatFromTSP4ForMetaModelling(QcPSPData, list_ID_PE_MES)
}
\arguments{
\item{QcPSPData}{the database that is retrieved through the restoreQcPSPData function}

\item{list_ID_PE_MES}{a vector of numerics that stand for the ID_PE_MES}
}
\value{
a data.frame object formatted for Capsis Web API
}
\description{
Extract plot list for Artemis-2009 simulation
}
