% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QcPSP.R
\name{getMetaData}
\alias{getMetaData}
\title{Provide the metadata for any of the seven data.frame objects}
\usage{
getMetaData(tableName)
}
\arguments{
\item{tableName}{a string among these six: plots, plotMeasurements, treeIndex,
treeMeasurements, saplings, and sbwDefoliations}
}
\value{
a data.frame object containing the metadata.
}
\description{
Provide the metadata for any of the seven data.frame objects
}
