% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QcPSP.R
\docType{data}
\name{measurementIndex}
\alias{measurementIndex}
\title{A data.frame object containing the plot:measurement information.}
\format{
An object of class \code{data.frame} with 50430 rows and 134 columns.
}
\usage{
data(measurementIndex)
}
\description{
A data.frame object containing the plot:measurement information.
}
\examples{
QcMeasurementIndex <- QcPSP::measurementIndex
}
\keyword{datasets}
