% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QcPSP.R
\docType{data}
\name{plotIndex}
\alias{plotIndex}
\title{A data.frame object containing the plot-level information.}
\format{
An object of class \code{data.frame} with 12816 rows and 29 columns.
}
\usage{
data(plotIndex)
}
\description{
A data.frame object containing the plot-level information.
}
\examples{
QcPlotIndex <- QcPSP::plotIndex
}
\keyword{datasets}
