% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QcPSP.R
\docType{data}
\name{QcMeasurementIndex}
\alias{QcMeasurementIndex}
\title{A data.frame object containing the plot:measurement information.}
\format{
An object of class \code{data.frame} with 50430 rows and 134 columns.
}
\usage{
data(QcMeasurementIndex)
}
\description{
A data.frame object containing the plot:measurement information.
}
\examples{
data(QcMeasurementIndex)
}
\keyword{datasets}
