% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QcPSP.R
\docType{data}
\name{QcTreeIndex}
\alias{QcTreeIndex}
\title{A data.frame object containing the tree-level information.}
\format{
An object of class \code{data.frame} with 814232 rows and 9 columns.
}
\usage{
data(QcTreeIndex)
}
\description{
A data.frame object containing the tree-level information.
}
\examples{
data(QcTreeIndex)
}
\keyword{datasets}
