% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QcPSP.R
\docType{data}
\name{QcTreeMeasurements}
\alias{QcTreeMeasurements}
\title{A data.frame object containing the tree:measurement information.}
\format{
An object of class \code{data.frame} with 2089280 rows and 13 columns.
}
\usage{
data(QcTreeMeasurements)
}
\description{
A data.frame object containing the tree:measurement information.
}
\examples{
data(QcTreeMeasurements)
}
\keyword{datasets}
