% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QcPSP.R
\name{restoreQcPSPData}
\alias{restoreQcPSPData}
\title{Restore Quebec PSP Data in the Global Environment.}
\usage{
restoreQcPSPData()
}
\description{
This function call creates a list with 7 data.frame objects.
}
\details{
The five data.frame objects are: \cr
\itemize{
\item plots: the index of the plots \cr
\item treeIndex: the index of the trees \cr
\item plotMeasurements: the index of the plot measurements \cr
\item treeMeasurements: the tree measurements \cr
\item saplings: the tally of the sapling in the 40 m2 subplot \cr
\item sbwDefoliations : the spruce budworm defoliation indices for the plot \cr
\item photoInterpretedStands: the photo interpreted stand around the plot \cr
}
}
